Encapsulate channel merge logic in update_channels()
